modem.py: Update login procedure for ZTE H1600
Logging in to ZTE H1600 firmware version V7.0.3_OTE.3.T7A was failing
with

---
Traceback (most recent call last):
  File "/home/vangelis/git/hub/eimparas/TR069-Bot/ModemBot/./simple.py", line 9, in <module>
    z.updateStats()
  File "/home/vangelis/git/hub/eimparas/TR069-Bot/ModemBot/modem.py", line 269, in updateStats
    if json.loads(response)["login_need_refresh"] :
       ~~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^
KeyError: 'login_need_refresh'
---

because of this HTTP error response:
{"lockingTime":-1,"loginErrMsg":"This page has expired, please refresh and try again. ","promptMsg":"","sess_token":"<redacted>"}

Capturing an HAR log from Forefox shows that the page the router itself
serves includes an "_sessionTOKENByPost" parameter in HTTP POST, so this
may have been a recent change in the firmware for the H1600.

Update the login procedure in modem.py to also include this parameter in
the HTTP POST call.

Refs #3

Signed-off-by: Vangelis Koukis <vkoukis@gmail.com>