modem.py: Handle login failure for ZTE H1600
Inspect the JSON response for the "loginErrMsg" field when logging in to
ZTE H1600 to handle login failures gracefully.

Before this change, attempting to log in with the wrong password led to
this exception:

---
Traceback (most recent call last):
  File "/home/vangelis/git/hub/vkoukis/eimparas-TR069-Bot/ModemBot/./simple.py", line 9, in <module>
    z.updateStats()
  File "/home/vangelis/git/hub/vkoukis/eimparas-TR069-Bot/ModemBot/modem.py", line 275, in updateStats
    if json.loads(response)["login_need_refresh"] :
       ~~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^
KeyError: 'login_need_refresh'
---

This enables a clearer error message, e.g., when attempting to log in
with wrong credentials:

---
$ python3 cli.py
Traceback (most recent call last):
  File "/home/vangelis/git/hub/vkoukis/eimparas-TR069-Bot/ModemBot/cli.py", line 9, in <module>
    z.updateStats()
  File "/home/vangelis/git/hub/vkoukis/eimparas-TR069-Bot/ModemBot/modem.py", line 277, in updateStats
    raise Exception("Failed to log in to ZTE H1600: %s" %
Exception: Failed to log in to ZTE H1600: Username or password is wrong.
---

Also include a small test program, cli.py, to instantiate and test the
Modem class directly from the CLI.

Closes #3

Signed-off-by: Vangelis Koukis <vkoukis@gmail.com>